Refactor the modal JS module
