+++ packages/modal/example.html
@@ -49,10 +49,10 @@
         <li>
           <button data-modal-open="modal-default" class="link">Modal Default</button>
           <div data-modal="modal-default" class="modal">
-            <div data-modal-dialog class="modal__dialog dialog">
+            <div data-modal-dialog class="modal__dialog dialog" role="dialog" aria-modal="true" aria-labelledby="modal-default-title" tabindex="-1">
               <button data-modal-close class="dialog__close link">Close</button>
               <div class="dialog__body">
-                <p>Modal Default</p>
+                <p id="modal-default-title">Modal Default</p>
               </div>
             </div>
           </div>
@@ -60,11 +60,11 @@
 
         <li>
           <button data-modal-open="modal-focus-trap" class="link">Modal Focus Trap</button>
-          <div data-modal="modal-focus-trap" class="modal is-opened">
-            <div data-modal-dialog class="modal__dialog dialog">
+          <div data-modal="modal-focus-trap" class="modal">
+            <div data-modal-dialog class="modal__dialog dialog" role="dialog" aria-modal="true" aria-labelledby="modal-focus-trap-title" tabindex="-1">
               <a class="display-none" href="#">Focusable, but hidden...</a>
               <div class="dialog__header">
-                <h2 class="dialog__title">Modal Focus Trap</h2>
+                <h2 id="modal-focus-trap-title" class="dialog__title">Modal Focus Trap</h2>
                 <button data-modal-close class="dialog__close link">Close</button>
               </div>
               <div class="dialog__body gap-y">
