+++ packages/modal/src/js/initialState.js
@@ -2,20 +2,23 @@ import { addClass, removeClass } from '@vrembem/core/index';
 import { setInert, setOverflowHidden } from '@vrembem/core/index';
 import { focusTrigger } from '@vrembem/core/index';
 
-export function setInitialState(obj) {
-  const modals = document.querySelectorAll(`[data-${obj.settings.dataModal}]`);
+export function setInitialState() {
+  const modals = document.querySelectorAll(`[data-${this.settings.dataModal}]`);
   modals.forEach((el) => {
-    if (el.classList.contains(obj.settings.stateOpened)) {
-      setInert(false, obj.settings.selectorInert);
-      setOverflowHidden(false, obj.settings.selectorOverflow);
-      focusTrigger(obj);
-      obj.focusTrap.destroy();
+    // Remove opened state setup
+    if (el.classList.contains(this.settings.stateOpened)) {
+      setInert(false, this.settings.selectorInert);
+      setOverflowHidden(false, this.settings.selectorOverflow);
+      focusTrigger(this);
+      this.focusTrap.destroy();
     }
+
+    // Remove all state classes and add the default state (closed)
     removeClass(el,
-      obj.settings.stateOpened,
-      obj.settings.stateOpening,
-      obj.settings.stateClosing
+      this.settings.stateOpened,
+      this.settings.stateOpening,
+      this.settings.stateClosing
     );
-    addClass(el, obj.settings.stateClosed);
+    addClass(el, this.settings.stateClosed);
   });
 }
